Added support for appengine Managed VMs by splitting up Controller struct declaration into two files, which are selectively compiled with the use of build constraints
